Change to source directory before iterating
This updates the start of the command to change to the source directory before looking for files, so that the for loop doesn't use the "source" variable. This strategy seemed to work on my machine.

Also updates the source and base_destination declarations so that the variables themselves don't have quotes inside them, otherwise they seem to get double-quoted.